backend/requirements.txt
Django==5.2.6
django-ninja==1.4.3
django-cors-headers==4.7.0
Pillow==11.3.0
djangorestframework==3.16.1
